@@ -1,19 +1,16 @@
 package at.haselwanter.android.lili_example.fragments;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import at.haselwanter.android.lili.adapters.ListAdapter;
 import at.haselwanter.android.lili.adapters.OneLineImageItemAdapter;
 import at.haselwanter.android.lili.fragments.ListFragment;
-import at.haselwanter.android.lili_example.ListActivity;
 import at.haselwanter.android.lili_example.models.OneLineDummy;
+import at.haselwanter.android.lili_example.viewmodels.OneLineDummyModel;
 
 
 /**
  * Created by Stefan Haselwanter on 14.09.2017.
  */
-public class OneItemFragment extends ListFragment<OneLineDummy> {
+public class OneItemFragment extends ListFragment<OneLineDummy, OneLineDummyModel> {
     @Override
     public String getFragmentTag() {
         return "one_item_fragment";
@@ -25,19 +22,11 @@ public class OneItemFragment extends ListFragment<OneLineDummy> {
     }
 
     @Override
-    public List<OneLineDummy> loadData() {
-        List<OneLineDummy> list = new ArrayList<>();
-
-        ListActivity.simulateWaitingForData();
-
-        for (int i = 0; i < ListActivity.NUMBER_OF_ITEMS; i++)
-            list.add(new OneLineDummy());
-
-        return list;
+    protected Class<OneLineDummyModel> getModelClass() {
+        return OneLineDummyModel.class;
     }
 
-    @Override
-    public void onNoDataLoaded() {
-
+    public void loadMoreData() {
+        model.loadDataAsync();
     }
 }
