@@ -7,7 +7,7 @@ import at.haselwanter.android.lili_example.R;
  * Created by Stefan Haselwanter on 15.09.2017.
  */
 public class OneLineDummy extends OneLineImageItem {
-    public OneLineDummy() {
-        super(0, "First line", R.drawable.ic_android);
+    public OneLineDummy(long id, String name) {
+        super(id, name, R.drawable.ic_android);
     }
 }
